+++ examples/genUUID.js
@@ -1,0 +1,8 @@
+const jssuper = require("jssuper");
+
+// EX 1
+// Generate a UUID v4
+// Output = "f47ac10b-58cc-4372-a567-0e02b2c3d479" Or another UUID
+let uuid = jssuper.genUUID();
+
+console.log(uuid);